[ANDROID] Refactoring
