+++ Native/Android/LunarConsole/lunarConsole/src/main/java/spacemadness/com/lunarconsole/console/BaseConsoleActionAdapter.java
@@ -21,8 +21,6 @@
 
 package spacemadness.com.lunarconsole.console;
 
-import spacemadness.com.lunarconsole.console.actions.IdentityEntry;
-
 public abstract class BaseConsoleActionAdapter<T extends IdentityEntry> extends BaseConsoleAdapter<T>
 {
     public BaseConsoleActionAdapter(DataSource<T> dataSource)
