+++ Native/Android/LunarConsole/lunarConsole/src/main/java/spacemadness/com/lunarconsole/console/VariableType.java
@@ -1,5 +1,5 @@
 //
-//  EntryType.java
+//  VariableType.java
 //
 //  Lunar Unity Mobile Console
 //  https://github.com/SpaceMadness/lunar-unity-console
@@ -19,15 +19,24 @@
 //  limitations under the License.
 //
 
-package spacemadness.com.lunarconsole.console.actions;
+package spacemadness.com.lunarconsole.console;
 
-/**
- * Created by alementuev on 1/23/17.
- */
+import spacemadness.com.lunarconsole.debug.Log;
 
-public enum EntryType
+public enum VariableType
 {
-    Action,
-    Variable,
-    Header
+    Unknown, Boolean, Integer, Float, String;
+
+    public static VariableType parse(String name)
+    {
+        try
+        {
+            return Enum.valueOf(VariableType.class, name);
+        }
+        catch (Exception e)
+        {
+            Log.e(e, "Exception while parsing variable type: %s", name);
+            return VariableType.Unknown;
+        }
+    }
 }
